[Add] IndicatorScreen for PatientCardScreen
